# Use fsspec.parquet module for better remote-storage read_parquet performance