--- a/dask/dataframe/io/parquet/arrow.py
+++ b/dask/dataframe/io/parquet/arrow.py
@@ -313,7 +313,7 @@ def _process_kwargs(kwargs):
     dataset_kwargs = user_kwargs.pop("dataset", {})
     read_kwargs = user_kwargs.pop("read", {})
     arrow_to_pandas_kwargs = user_kwargs.pop("arrow_to_pandas", {})
-    if "open_parquet_file" not in user_kwargs:
+    if "open_parquet_file" in user_kwargs:
         # Allow user to pass "open_parquet_file"
         # outside of the "read" kwargs
         read_kwargs["open_parquet_file"] = user_kwargs.pop("open_parquet_file", {})