Depends on [fsspec#813](https://github.com/intake/filesystem_spec/pull/813)

The new `fsspec.parquet.open_parquet_file` function offers an optimized version of `fs.open` for parquet-formatted data files. This PR updates both the "pyarrow" and "fastparquet" `read_parquet` implementations to (optionally) leverage this new feature.

**API**:

The **new** API change proposed in this PR is to add an `open_file_options` kwarg to `read_parquet`. This kwarg is intended as a central dictionary location for the user to define all options related to data-file opening.  Within this dictionary, the user can specify an optimized "parquet" precaching strategy within a nested `"cache_options"` dictionary:

```python
ddf = dd.read_parquet(paths, ..., open_file_options={"cache_options": {"precache": "parquet"}})
```

We choose to define all cache-related options under `"cache_options"` to simplify current and future fsspec integration.  Although fsspec does not yet support the `"precache"` sub-field directly, there seems to be agreement that such a key can be recognized in the future (within `fs.open` and related).  Until then, Dask can simply perform a manual check for `{"precache": "parquet"}` and redirect the `fs.open` call to `open_parquet_file`.

**Blockers**:

- [x] Merge/release [fsspec#813](https://github.com/intake/filesystem_spec/pull/813) (or check that CI passes with `master` fsspec branch)
- [x] Ensure that the ability to `import fsspec.parquet` is a sufficient check for compatibility (or improve this check if it is not)
- [x] Add test to confirm that passing an `open_parquet_file={}` kwarg to `read_parquet` can be used to pass kwargs to `fsspec.parquet.open_parquet_file`

**Performance Motivation**

```python
import dask.dataframe as dd

%%time
df = dd.read_parquet(
    ["gs://my-bucket/criteo-parquet/day_0.parquet"],
    columns=["I10"],
    split_row_groups=10,
    engine=...,
    storage_options=...,
    open_file_options={"cache_options": {"precache": "parquet"}},
)
```
```
# Main Branch - "fastparquet" Engine
CPU times: user 6.65 s, sys: 5 s, total: 11.6 s
Wall time: 8.52 s

# Main Branch - "pyarrow" Engine
CPU times: user 6.83 s, sys: 4.75 s, total: 11.6 s
Wall time: 8.04 s

# This PR - "fastparquet" Engine
CPU times: user 5.51 s, sys: 3.7 s, total: 9.21 s
Wall time: 4.78 s

# This PR - "pyarrow" Engine
CPU times: user 5.37 s, sys: 3.7 s, total: 9.07 s
Wall time: 4.75 s
```