@jorisvandenbossche 2021-11-09
I am not fully sure if the pyarrow-based engine should (always) make use of this. Similar optimization are already done by the parquet/dataset readers, and so if we don't get benefit of that when used through dask (which seems to be the case based on your timings above), we should maybe rather investigate why this is not the case.