@rjzamora 2021-11-10
>I am not fully sure if the pyarrow-based engine should (always) make use of this. Similar optimization are already done by the parquet/dataset readers, and so if we don't get benefit of that when used through dask (which seems to be the case based on your timings above), we should maybe rather investigate why this is not the case.

Yes - There is no reason that pyarrow cannot optimize the data transfer in a similar way, and so I do want the final iteration of this to make the `fsspec.parquet.open_parquet_file` approach to be optional (with the default perhaps depending on the pyarrow version).