@rjzamora 2021-12-13
>I am pretty busy in company things this week, but I'll get to it

Understandable - The only pressing thing about that PR is the decision of "how" to expose the file-opening options to the user. Right now its in a `open_options={"file_format": "parquet", ...}` kwarg.  There is pressure to use the `fsspec.parquet` optimization in RAPIDS/NVTabular, so the decision will be made in one of those libraries if we don't make it in Dask first :)

>Any chance #8470 can precede this one?

Sure! I'll prioritize it for today.  Sorry - Just seeing that PR now (my notifications are a mess right now).