@martindurant 2021-12-13
> There is pressure to use the fsspec.parquet optimization in RAPIDS/NVTabular, so the decision will be made in one of those libraries if we don't make it in Dask first

That's fine. I am a little surprised, though - I had pictures the parquet module going out with the release and having a longer period of general community feedback following the blog post.