@martindurant 2021-12-13
I think I am persuaded to push ahead, since then we get RAPIDS users too, and maybe more eyes on the fsspec code :)