@rjzamora 2021-12-14
@martindurant - I know it will take time before you can give this a careful review. That is totally fine - No rush on my end. 

The only question I’d like to get your thoughts on “sooner” (rather than later) is the **public** change to the `read_parquet` API.  Right now, this PR proposes a new `open_options` parameter, where the user can pass in a dictionary of key-word arguments for a new `dd.io.utils.open_input_files` function. Within this dictionary, the user is free to specify a specific “file_format” item to turn on/off format-optimized file-opening behavior. Therefore, for `read_parquet`, `{“file_format”: “parquet”}` will be added to `open_options` automatically if the user does not specify `{“file_format”: None}` (or `{“open_file_cb”: …}` for a custom-open function).

  Therefore, if a user want s to avoid using `fsspec.parquet`, they would need to do something like: `dd.read_parquet(path, open_options={“file_format”: None})`.  If they also wanted to use a non-default caching strategy, it would be something like:  `dd.read_parquet(path, open_options={“file_format”: None, “cache_type”: “all”})`.
  
  Does this seem reasonable to you? Does the `open_options` dictionary and/or the `"file_format"` key make sense?  I am happy to revise and discuss.