@martindurant 2021-12-16
To the specific API:

- I think it's fine to add the extra kwarg. The term `open_options` is ambiguous, though, because the dataset is also "opened". Maybe `file_open_options`
- adding `open_input_files` is fine, but in practice I don't expect any user to call it or know about it, and I don't expect the "open_file_cb" argument to be used either. The docs of what this all means probably go elsewhere
- `{"file_format": "parquet"}` should still be considered experimental and not the default? It's also a bit confusing in the name, since indeed the format is parquet, here we are saying that we want to use parquet-specific logic for our file access pattern (i.e., precaching). It *feels* more like an fsspec storage_option. Maybe fsspec's open functions should allow for an `open_options` that gets passed to the `fs.open` calls? That latter idea is definitely a good idea, on reflection, but I'm not sure if it helps simplify the code here at all (maybe stuff the rg, cols into cache_options... etc).